+++ techbuzz/src/main/resources/templates/users/registrationStatus.html
@@ -12,6 +12,10 @@
 
     <div class="col-md-4 offset-4 text-center">
         <div class="card">
+            <div class="alert alert-success" role="alert"  th:if="${message}">
+                <span th:text="${message}"></span>
+                <button aria-label="Close" class="btn-close" data-bs-dismiss="alert" type="button"></button>
+            </div>
             <div class="card-body">
                 <div class="card-title">
                     <h4>Registration is successful</h4>
